fix(api): revert ConnectedPeerResponse fields to u128 with suffix on schema examples
u128 is the native type from Duration::as_millis(). The _u128 suffix is
only needed in the #[schema(example)] attribute to avoid i32 overflow at
compile time — it does not appear in the JSON sent to clients.

+++ hoprd/rest-api/src/network.rs
@@ -104,7 +104,7 @@ pub(super) async fn probability(State(state): State<Arc<InternalState>>) -> impl
 #[schema(example = json!({
     "address": "0xb4ce7e6e36ac8b01a974725d5ba730af2b156fbe",
     "probeRate": 0.476,
-    "lastUpdate": 1690000000000_u64,
+    "lastUpdate": 1690000000000_u128,
     "averageLatency": 100,
     "score": 0.7
 }))]
@@ -116,11 +116,11 @@ pub(crate) struct ConnectedPeerResponse {
     #[schema(example = 0.476)]
     probe_rate: f64,
     /// Epoch milliseconds of the last observation update.
-    #[schema(example = 1690000000000_u64)]
-    last_update: u64,
+    #[schema(example = 1690000000000_u128)]
+    last_update: u128,
     /// Average latency in milliseconds, if available.
     #[schema(example = 100)]
-    average_latency: Option<u64>,
+    average_latency: Option<u128>,
     #[schema(example = 0.7)]
     score: f64,
 }
@@ -172,8 +172,8 @@ pub(super) async fn connected(State(state): State<Arc<InternalState>>) -> impl I
         peers.push(ConnectedPeerResponse {
             address,
             probe_rate: imm.average_probe_rate(),
-            last_update: obs.last_update().as_millis() as u64,
-            average_latency: imm.average_latency().map(|l| l.as_millis() as u64),
+            last_update: obs.last_update().as_millis(),
+            average_latency: imm.average_latency().map(|l| l.as_millis()),
             score: obs.score(),
         });
     }
